cambio de tipo de datos y controlprueba

+++ vehiculos-service/src/main/java/org/backend/entities/Empleado.java
@@ -24,6 +24,6 @@ public class Empleado {
     private String apellido;
 
     @Column(name = "TELEFONO_CONTACTO", nullable = false)
-    private int telefono_contacto;
+    private Long telefono_contacto;
 
 }
